examples/blog/requirements.txt
anom
awesome-slugify
bottle
markdown
passlib
